Building and Training Pytorch model

+++ nltk_utils.py
@@ -3,7 +3,6 @@ import numpy as np
 # nltk.download('punkt')
 from nltk.stem.porter import PorterStemmer
 
-
 def tokenize(sentence):
     return nltk.word_tokenize(sentence)
 
@@ -13,12 +12,9 @@ def stem(word):
 
 def create_bag_of_words(sentence_tokens,all_words):
     sentence_tokens = [stem(w) for w in sentence_tokens]
-    bag_of_words = np.zeros(len(all_words),dtype=np.int32)
+    bag_of_words = np.zeros(len(all_words),dtype=np.float64)
     for idx,word in enumerate(all_words):
         if word in sentence_tokens:
-            bag_of_words[idx] = 1
+            bag_of_words[idx] = 1.0
     return bag_of_words
-
-
-
